Move smoothing to preprocessing
