+++ skfda/preprocessing/__init__.py
@@ -1,1 +1,2 @@
 from . import registration
+from . import smoothing
