+++ skfda/preprocessing/smoothing/__init__.py
@@ -1,0 +1,2 @@
+from . import kernel_smoothers
+from . import validation
