+++ skfda/preprocessing/smoothing/validation.py
